Remove `BUNDLE_BIN`
`BUNDLE_BIN` is a path that is supposed to be specific to a given
bundle. As such it cannot and should not be reused.

This manifests as bundle-specific path entries being hardcoded in
scripts of a potentially shared `bin` folder. As a consequence the
scripts can fail unless paths happen to be identical by sheer
coincidence.