Address musl failure
We have a complcated scenario:

When the single-arch image is exported for tagging locally it's only a
single-arch image, which makes the multi-arch build fail because there's
no FROM that can satisfy the other arch.

So we remove the local tag, but then it means the FROM references a
remote image; but if it doesn't exist then the process fails for both
arches; and when it does then the FROM references the remote image, not
the one that has just been built.

As a workaround we only push the commit image when we just pushed the
dependency image.